# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Bulk Purchasing Platform API
  description: |
    Complete API documentation for the Save Together bulk purchasing application.
    
    ## Features
    - User authentication with JWT tokens
    - Product catalog management
    - Vendor management and dashboards
    - Bulk purchasing groups
    - Order management with MOQ validation
    - Escrow-protected transactions
    - Dispute resolution
    - Real-time chat with WebSocket (Socket.IO)
    
    ## Authentication
    Most endpoints require a JWT token obtained from `/auth/login`. Include the token in the Authorization header:
    ```
    Authorization: Bearer {token}
    ```
  version: 1.0.0
  contact:
    name: API Support
    email: support@savetogether.com
  license:
    name: MIT
servers:
  - url: http://localhost:3001/api
    description: Development Server
  - url: https://api.savetogether.com
    description: Production Server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from login endpoint. Format is "Bearer {token}"
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [superUser, admin, vendor, member]
          description: User's global role in the system
        avatar:
          type: string
          format: uri
          nullable: true
        vendor_id:
          type: string
          format: uuid
          nullable: true
          description: ID of vendor if role is 'vendor'
        trust_score:
          type: integer
          nullable: true
        is_active:
          type: boolean
        is_online:
          type: boolean
        created_at:
          type: string
          format: date-time
      required: [id, email, name, role, created_at]
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
      required: [success, error]
...
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List all users
      security:
        - bearerAuth: []
      description: Admin only. Get all users in the system
      responses:
        '200':
          description: Users list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'

    post:
      tags: [Users]
      summary: Create user
      security:
        - bearerAuth: []
      description: Admin only. Create new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 6
                name:
                  type: string
                role:
                  type: string
                  enum: [superUser, admin, vendor, member]
                vendor_id:
                  type: string
                  format: uuid
                  nullable: true
              required: [email, password, name, role]
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/User'
  /users/{id}:
    get:
      tags: [Users]
      summary: Get user by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/User'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags: [Users]
      summary: Update user
      security:
        - bearerAuth: []
      description: Admin only. Update user details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                role:
                  type: string
                  enum: [superUser, admin, vendor, member]
                vendor_id:
                  type: string
                  format: uuid
                  nullable: true
                is_active:
                  type: boolean
                trust_score:
                  type: integer
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/User'

    delete:
      tags: [Users]
      summary: Delete user
      security:
        - bearerAuth: []
      description: SuperUser only. Permanently delete user account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
  /users/{id}/activate:
    post:
      tags: [Users]
      summary: Activate user
      security:
        - bearerAuth: []
      description: Admin only. Re-activate deactivated user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User activated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/User'
  /users/{id}/deactivate:
    post:
      tags: [Users]
      summary: Deactivate user
      security:
        - bearerAuth: []
      description: Admin only. Disable user account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User deactivated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/User'
  /users/stats:
    get:
      tags: [Users]
      summary: User statistics
      security:
        - bearerAuth: []
      description: Admin only. Get user system statistics
      responses:
        '200':
          description: User statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      total:
                        type: integer
                      active:
                        type: integer
                      superUsers:
                        type: integer
                      admins:
                        type: integer
                      vendors:
                        type: integer
                      members:
                        type: integer

  # ============ PRODUCTS ============
...